🎨 Split session and flushEvent into seperate files
